requirements.txt.txt
pandas==2.1.0
PyMySQL==1.0.3
SQLAlchemy==2.0.20
streamlit==1.26.0
